PWN-7171 - New Swap. Fix comments
