+++ app/src/main/java/org/p2p/wallet/swap/jupiter/statemanager/SwapCoroutineScope.kt
@@ -1,0 +1,10 @@
+package org.p2p.wallet.swap.jupiter.statemanager
+
+import kotlin.coroutines.CoroutineContext
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
+
+class SwapCoroutineScope : CoroutineScope {
+    override val coroutineContext: CoroutineContext = SupervisorJob() + Dispatchers.IO
+}
