+++ app/src/main/java/org/p2p/wallet/swap/jupiter/statemanager/SwapStateHandler.kt
@@ -1,8 +1,8 @@
 package org.p2p.wallet.swap.jupiter.statemanager
 
-import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
 
 interface SwapStateHandler {
     fun canHandle(oldState: SwapState): Boolean
-    fun handle(oldState: SwapState, action: SwapStateAction): Flow<SwapState>
+    suspend fun handleAction(state: MutableStateFlow<SwapState>, action: SwapStateAction)
 }
