+++ app/src/main/java/org/p2p/wallet/swap/jupiter/statemanager/SwapStateTokenAZeroHandler.kt
@@ -1,59 +1,38 @@
 package org.p2p.wallet.swap.jupiter.statemanager
 
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.emitAll
-import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.MutableStateFlow
 
 class SwapStateTokenAZeroHandler(
-    private val swapRoutesRefresher: SwapRoutesRefresher
+    private val swapRoutesRefresher: SwapStateRoutesRefresher
 ) : SwapStateHandler {
 
     override fun canHandle(oldState: SwapState): Boolean = oldState is SwapState.TokenAZero
 
-    override fun handle(oldState: SwapState, action: SwapStateAction): Flow<SwapState> {
+    override suspend fun handleAction(state: MutableStateFlow<SwapState>, action: SwapStateAction) {
+        val oldState = state.value
         oldState as SwapState.TokenAZero
-        return flow {
-            when (action) {
-                is SwapStateAction.SlippageChanged -> {
-                    emit(
-                        SwapState.TokenAZero(
-                            tokenA = oldState.tokenA,
-                            tokenB = oldState.tokenB,
-                            slippage = action.newSlippageValue
-                        )
-                    )
-                }
-                is SwapStateAction.TokenAAmountChanged -> {
-                    emitAll(
-                        swapRoutesRefresher.refreshRoutes(
-                            tokenA = oldState.tokenA,
-                            tokenB = oldState.tokenB,
-                            amountTokenA = action.newAmount,
-                            slippage = oldState.slippage,
-                            activeRouteOrdinal = 1
-                        )
-                    )
-                }
-                is SwapStateAction.TokenAChanged -> {
-                    emit(
-                        SwapState.TokenAZero(
-                            tokenA = action.newTokenA,
-                            tokenB = oldState.tokenB,
-                            slippage = oldState.slippage
-                        )
-                    )
-                }
-                is SwapStateAction.TokenBChanged -> {
-                    emit(
-                        SwapState.TokenAZero(
-                            tokenA = oldState.tokenA,
-                            tokenB = action.newTokenB,
-                            slippage = oldState.slippage
-                        )
-                    )
-                }
-                else -> Unit
+
+        when (action) {
+            is SwapStateAction.SlippageChanged -> {
+                state.value = oldState.copy(slippage = action.newSlippageValue)
             }
+            is SwapStateAction.TokenAAmountChanged -> {
+                swapRoutesRefresher.refreshRoutes(
+                    state = state,
+                    tokenA = oldState.tokenA,
+                    tokenB = oldState.tokenB,
+                    amountTokenA = action.newAmount,
+                    slippage = oldState.slippage,
+                    activeRouteOrdinal = 1
+                )
+            }
+            is SwapStateAction.TokenAChanged -> {
+                state.value = oldState.copy(tokenA = action.newTokenA)
+            }
+            is SwapStateAction.TokenBChanged -> {
+                state.value = oldState.copy(tokenB = action.newTokenB)
+            }
+            else -> Unit
         }
     }
 }
